Reorganise component folder structure
