+++ resources/js/components/resources/PublishActions.vue
