+++ resources/js/components/resources/PublishForm.vue
